requirements.txt
streamlit==1.37.1
pytesseract==0.3.10
pillow==10.4.0